Reworked localization
Localization is now able to be viewed in Localizable.xcstrings. App texts have been reworked to use this string catalog instead of language switches. Reused titles will use the same key for the localization (e.g. about.privacyPolicy for both the button in the "About" screen and as a title for the actual info page).
Bold text was removed in this commit, will be reimplemented in the next